requirements.txt
discord.py
python-dotenv
nest_asyncio
redis
requests
selenium
